\name{@GMAPOBJNAME@}
\docType{package}

\alias{@PKGNAME@-package}
\alias{@PKGNAME@}
\alias{@GMAPOBJNAME@}


\title{@PKGTITLE@}

\description{  
  This package loads a GmapGenome object @PKGNAME@. Such objects are
  an means to distribute and share genomes and associated SNPs and
  slices for use with GSNAP, GMAP, and bam_tally.
}

\author{@AUTHOR@}


\seealso{
  \link[gmapR]{GmapGenome},
}

\examples{
## load the library
library(@PKGNAME@)
## list the contents that are loaded into memory
ls('package:@PKGNAME@')
## show the GmapGenome object that is loaded by calling it's name
@PKGNAME@
}

\keyword{package}
\keyword{data}
